+++ packages/solutions/module-tools/src/types/config/index.ts
@@ -1,7 +1,6 @@
 import type { BuildOptions } from 'esbuild';
 import type { CreateFilter } from '@rollup/pluginutils';
 import type { MinifyOptions as TerserMinifyOptions } from 'terser';
-import type { TestConfig } from '@modern-js/types';
 import type { Config } from '../../../compiled/@svgr/core';
 import { internalPreset, presetList } from '../../constants/preset';
 import { ICompiler } from '../esbuild';
@@ -238,8 +237,6 @@ export interface ModuleExtraConfig {
   buildPreset?: BuildPreset;
 
   dev?: Dev;
-
-  testing?: TestConfig;
 
   runtime?: RuntimeUserConfig;
 }