+++ packages/cli/plugin-tailwind/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @modern-js/plugin-tailwindcss
 
+## 2.52.0
+
+### Patch Changes
+
+- Updated dependencies [85ac453]
+  - @modern-js/runtime@2.52.0
+  - @modern-js/node-bundle-require@2.52.0
+  - @modern-js/utils@2.52.0
+
 ## 2.51.0
 
 ### Patch Changes