+++ packages/toolkit/runtime-utils/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @modern-js/runtime-utils
 
+## 2.52.0
+
+### Minor Changes
+
+- 85ac453: feat: support control status code by data loader
+  feat: 支持通过 data loader 控制页面状态码
+
+### Patch Changes
+
+- @modern-js/utils@2.52.0
+
 ## 2.51.0
 
 ### Patch Changes